requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.51.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
httpie==3.2.1
idna==3.3
Markdown==3.3.7
multidict==6.0.2
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.4
sqlparse==0.4.2
urllib3==1.26.9